+++ FabrikaLine_Backend/src/main/java/com/example/fabrikaline_backend/Services/OrderStatusServiceImpl.java
@@ -58,7 +58,7 @@ public class OrderStatusServiceImpl implements IOrderStatusService, IAbstractSer
     }
 
     @Override
-    public OrderStatus getById(int id) {
+    public OrderStatus getById(Long id) {
         return null; //TODO
     }
 }
